+++ tests/branching_kernels/if_else_1_label/if_else_1_label_hands.cl
@@ -3,9 +3,9 @@ void if_else_1_label(int x, __global int *data, int y)
 {
     uint *var2;
     uint var4;
-    if (2 == get_global_id(0)) {
+    if ((int)2 == (int)get_global_id(0)) {
         var2 = &data[(get_global_id(0) * 4) / 4];
-        var4 = (get_global_id(1) * x) - y;
+        var4 = (ulong)(get_global_id(1) * x) - (ulong)y;
     }
     else {
         var2 = &data[(get_global_id(0) * 4) / 4];
