+++ tests/linear_kernels/multiplication/multiplication_hands.cl
@@ -157,9 +157,9 @@ void mul_get_global_offset_get_local_id(int x, __global int *data)
 __kernel __attribute__((reqd_work_group_size(8, 4, 2)))
 void mul_get_local_id_get_local_id(int x, __global int *data)
 {
-    data[(get_global_id(0) * 4) / 4] = get_local_id(0) * get_local_id(0);
-    data[(get_global_id(1) * 4) / 4] = get_local_id(1) * get_local_id(1);
-    data[(get_global_id(2) * 4) / 4] = get_local_id(2) * get_local_id(2);
+    data[(get_global_id(0) * 4) / 4] = (int)get_local_id(0) * (int)get_local_id(0);
+    data[(get_global_id(1) * 4) / 4] = (int)get_local_id(1) * (int)get_local_id(1);
+    data[(get_global_id(2) * 4) / 4] = (int)get_local_id(2) * (int)get_local_id(2);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 16, 2)))
@@ -657,7 +657,7 @@ void mul_get_global_offset_64(int x, __global uint *data)
 __kernel __attribute__((reqd_work_group_size(64, 1, 1)))
 void mul_get_local_id_64(int x, __global uint *data)
 {
-    data[(get_global_id(0) * 4) / 4] = get_local_id(0) * get_local_id(0);
+    data[(get_global_id(0) * 4) / 4] = (int)get_local_id(0) * (int)get_local_id(0);
 }
 
 __kernel __attribute__((reqd_work_group_size(64, 1, 1)))
@@ -711,7 +711,7 @@ void mul_get_global_offset_8_8(int x, __global uint *data)
 __kernel __attribute__((reqd_work_group_size(8, 8, 1)))
 void mul_get_local_id_8_8(int x, __global uint *data)
 {
-    data[(get_global_id(0) * 4) / 4] = get_local_id(1) * get_local_id(0);
+    data[(get_global_id(0) * 4) / 4] = (int)get_local_id(1) * (int)get_local_id(0);
 }
 
 __kernel __attribute__((reqd_work_group_size(8, 8, 1)))
@@ -1077,7 +1077,7 @@ void mul_long_get_local_size_get_local_size(int x, __global ulong *data)
 __kernel __attribute__((reqd_work_group_size(2, 2, 4)))
 void mul_long_get_global_id_get_local_size(int x, __global ulong *data)
 {
-    data[((get_global_id(0) - get_global_offset(0)) * 8) / 8] = (((get_global_id(0) - get_global_offset(0)) + get_global_offset(0)) * 2) * 2;
+    data[((get_global_id(0) - get_global_offset(0)) * 8) / 8] = (((ulong)(get_global_id(0) - get_global_offset(0)) + (ulong)get_global_offset(0)) * 2) * 2;
     data[(get_global_id(1) * 8) / 8] = get_global_id(1) * 2;
     data[(get_global_id(2) * 8) / 8] = get_global_id(2) * 4;
 }
@@ -1109,72 +1109,72 @@ void mul_long_get_work_dim_get_local_size(int x, __global ulong *data)
 __kernel __attribute__((reqd_work_group_size(16, 2, 2)))
 void mul_long_x_get_global_id(__global long *data, int x)
 {
-    data[(get_global_id(0) * 8) / 8] = ((get_global_id(0) - get_global_offset(0)) + get_global_offset(0)) * x;
-    data[(get_global_id(1) * 8) / 8] = ((get_global_id(1) - get_global_offset(1)) + get_global_offset(1)) * x;
+    data[(get_global_id(0) * 8) / 8] = ((ulong)(get_global_id(0) - get_global_offset(0)) + (ulong)get_global_offset(0)) * x;
+    data[(get_global_id(1) * 8) / 8] = ((ulong)(get_global_id(1) - get_global_offset(1)) + (ulong)get_global_offset(1)) * x;
     data[(get_global_id(2) * 8) / 8] = get_global_id(2) * x;
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 16, 2)))
 void mul_long_get_global_offset_get_global_id(int x, __global long *data)
 {
-    data[(get_global_id(0) * 8) / 8] = ((get_global_id(0) - get_global_offset(0)) + get_global_offset(0)) * get_global_offset(0);
-    data[(get_global_id(1) * 8) / 8] = ((get_global_id(1) - get_global_offset(1)) + get_global_offset(1)) * get_global_offset(1);
+    data[(get_global_id(0) * 8) / 8] = ((ulong)(get_global_id(0) - get_global_offset(0)) + (ulong)get_global_offset(0)) * get_global_offset(0);
+    data[(get_global_id(1) * 8) / 8] = ((ulong)(get_global_id(1) - get_global_offset(1)) + (ulong)get_global_offset(1)) * get_global_offset(1);
     data[(get_global_id(2) * 8) / 8] = get_global_id(2) * get_global_offset(2);
 }
 
 __kernel __attribute__((reqd_work_group_size(8, 4, 2)))
 void mul_long_get_local_id_get_global_id(int x, __global long *data)
 {
-    data[(get_global_id(0) * 8) / 8] = get_local_id(0) * ((get_global_id(0) - get_global_offset(0)) + get_global_offset(0));
-    data[(get_global_id(1) * 8) / 8] = get_local_id(1) * ((get_global_id(1) - get_global_offset(1)) + get_global_offset(1));
+    data[(get_global_id(0) * 8) / 8] = get_local_id(0) * ((ulong)(get_global_id(0) - get_global_offset(0)) + (ulong)get_global_offset(0));
+    data[(get_global_id(1) * 8) / 8] = get_local_id(1) * ((ulong)(get_global_id(1) - get_global_offset(1)) + (ulong)get_global_offset(1));
     data[(get_global_id(2) * 8) / 8] = get_local_id(2) * get_global_id(2);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 16, 2)))
 void mul_long_get_group_id_get_global_id(int x, __global long *data)
 {
-    data[(get_global_id(0) * 8) / 8] = get_group_id(0) * ((get_global_id(0) - get_global_offset(0)) + get_global_offset(0));
-    data[(get_global_id(1) * 8) / 8] = get_group_id(1) * ((get_global_id(1) - get_global_offset(1)) + get_global_offset(1));
+    data[(get_global_id(0) * 8) / 8] = get_group_id(0) * ((ulong)(get_global_id(0) - get_global_offset(0)) + (ulong)get_global_offset(0));
+    data[(get_global_id(1) * 8) / 8] = get_group_id(1) * ((ulong)(get_global_id(1) - get_global_offset(1)) + (ulong)get_global_offset(1));
     data[(get_global_id(2) * 8) / 8] = get_group_id(2) * get_global_id(2);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 2, 4)))
 void mul_long_get_local_size_get_global_id(int x, __global ulong *data)
 {
-    data[(get_global_id(0) * 8) / 8] = ((get_global_id(0) - get_global_offset(0)) + get_global_offset(0)) * 2;
-    data[(get_global_id(1) * 8) / 8] = ((get_global_id(1) - get_global_offset(1)) + get_global_offset(1)) * 2;
+    data[(get_global_id(0) * 8) / 8] = ((ulong)(get_global_id(0) - get_global_offset(0)) + (ulong)get_global_offset(0)) * 2;
+    data[(get_global_id(1) * 8) / 8] = ((ulong)(get_global_id(1) - get_global_offset(1)) + (ulong)get_global_offset(1)) * 2;
     data[(get_global_id(2) * 8) / 8] = get_global_id(2) * 4;
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 2, 4)))
 void mul_long_get_global_id_get_global_id(int x, __global ulong *data)
 {
-    data[((get_global_id(0) - get_global_offset(0)) * 8) / 8] = get_global_id(0) * ((get_global_id(0) - get_global_offset(0)) + get_global_offset(0));
-    data[(get_global_id(1) * 8) / 8] = get_global_id(1) * ((get_global_id(1) - get_global_offset(1)) + get_global_offset(1));
+    data[((get_global_id(0) - get_global_offset(0)) * 8) / 8] = get_global_id(0) * ((ulong)(get_global_id(0) - get_global_offset(0)) + (ulong)get_global_offset(0));
+    data[(get_global_id(1) * 8) / 8] = get_global_id(1) * ((ulong)(get_global_id(1) - get_global_offset(1)) + (ulong)get_global_offset(1));
     data[(get_global_id(2) * 8) / 8] = get_global_id(2) * get_global_id(2);
 }
 
 __kernel __attribute__((reqd_work_group_size(16, 2, 2)))
 void mul_long_get_global_size_get_global_id(__global long *data, int x)
 {
-    data[(get_global_id(0) * 8) / 8] = get_global_size(0) * ((get_global_id(0) - get_global_offset(0)) + get_global_offset(0));
-    data[(get_global_id(1) * 8) / 8] = get_global_size(1) * ((get_global_id(1) - get_global_offset(1)) + get_global_offset(1));
+    data[(get_global_id(0) * 8) / 8] = get_global_size(0) * ((ulong)(get_global_id(0) - get_global_offset(0)) + (ulong)get_global_offset(0));
+    data[(get_global_id(1) * 8) / 8] = get_global_size(1) * ((ulong)(get_global_id(1) - get_global_offset(1)) + (ulong)get_global_offset(1));
     data[(get_global_id(2) * 8) / 8] = get_global_size(2) * get_global_id(2);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 2, 4)))
 void mul_long_get_num_groups_get_global_id(int x, __global ulong *data)
 {
-    data[(get_global_id(0) * 8) / 8] = get_num_groups(0) * ((get_global_id(0) - get_global_offset(0)) + get_global_offset(0));
-    data[(get_global_id(1) * 8) / 8] = get_num_groups(1) * ((get_global_id(1) - get_global_offset(1)) + get_global_offset(1));
+    data[(get_global_id(0) * 8) / 8] = get_num_groups(0) * ((ulong)(get_global_id(0) - get_global_offset(0)) + (ulong)get_global_offset(0));
+    data[(get_global_id(1) * 8) / 8] = get_num_groups(1) * ((ulong)(get_global_id(1) - get_global_offset(1)) + (ulong)get_global_offset(1));
     data[(get_global_id(2) * 8) / 8] = get_num_groups(2) * get_global_id(2);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 2, 4)))
 void mul_long_get_work_dim_get_global_id(int x, __global ulong *data)
 {
-    data[(get_global_id(0) * 8) / 8] = get_work_dim() * ((get_global_id(0) - get_global_offset(0)) + get_global_offset(0));
-    data[(get_global_id(1) * 8) / 8] = get_work_dim() * ((get_global_id(1) - get_global_offset(1)) + get_global_offset(1));
+    data[(get_global_id(0) * 8) / 8] = get_work_dim() * ((ulong)(get_global_id(0) - get_global_offset(0)) + (ulong)get_global_offset(0));
+    data[(get_global_id(1) * 8) / 8] = get_work_dim() * ((ulong)(get_global_id(1) - get_global_offset(1)) + (ulong)get_global_offset(1));
     data[(get_global_id(2) * 8) / 8] = get_work_dim() * get_global_id(2);
 }
 
@@ -1409,7 +1409,7 @@ void mul_long_get_global_offset_64(int x, __global ulong *data)
 __kernel __attribute__((reqd_work_group_size(64, 1, 1)))
 void mul_long_get_local_id_64(int x, __global ulong *data)
 {
-    data[(get_global_id(0) * 8) / 8] = get_local_id(0) * get_local_id(0);
+    data[(get_global_id(0) * 8) / 8] = (int)get_local_id(0) * (int)get_local_id(0);
 }
 
 __kernel __attribute__((reqd_work_group_size(64, 1, 1)))
@@ -1463,7 +1463,7 @@ void mul_long_get_global_offset_8_8(int x, __global ulong *data)
 __kernel __attribute__((reqd_work_group_size(8, 8, 1)))
 void mul_long_get_local_id_8_8(int x, __global ulong *data)
 {
-    data[(get_global_id(0) * 8) / 8] = get_local_id(1) * get_local_id(0);
+    data[(get_global_id(0) * 8) / 8] = (int)get_local_id(1) * (int)get_local_id(0);
 }
 
 __kernel __attribute__((reqd_work_group_size(8, 8, 1)))
@@ -1661,9 +1661,9 @@ void mul_char_get_global_offset_get_local_id(int x, __global char *data)
 __kernel __attribute__((reqd_work_group_size(8, 4, 2)))
 void mul_char_get_local_id_get_local_id(int x, __global char *data)
 {
-    data[get_global_id(0)] = get_local_id(0) * get_local_id(0);
-    data[get_global_id(1)] = get_local_id(1) * get_local_id(1);
-    data[get_global_id(2)] = get_local_id(2) * get_local_id(2);
+    data[get_global_id(0)] = (int)get_local_id(0) * (int)get_local_id(0);
+    data[get_global_id(1)] = (int)get_local_id(1) * (int)get_local_id(1);
+    data[get_global_id(2)] = (int)get_local_id(2) * (int)get_local_id(2);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 16, 2)))
@@ -2161,7 +2161,7 @@ void mul_char_get_global_offset_64(int x, __global uchar *data)
 __kernel __attribute__((reqd_work_group_size(64, 1, 1)))
 void mul_char_get_local_id_64(int x, __global uchar *data)
 {
-    data[get_global_id(0)] = get_local_id(0) * get_local_id(0);
+    data[get_global_id(0)] = (int)get_local_id(0) * (int)get_local_id(0);
 }
 
 __kernel __attribute__((reqd_work_group_size(64, 1, 1)))
@@ -2215,7 +2215,7 @@ void mul_char_get_global_offset_8_8(int x, __global uchar *data)
 __kernel __attribute__((reqd_work_group_size(8, 8, 1)))
 void mul_char_get_local_id_8_8(int x, __global uchar *data)
 {
-    data[get_global_id(0)] = get_local_id(1) * get_local_id(0);
+    data[get_global_id(0)] = (int)get_local_id(1) * (int)get_local_id(0);
 }
 
 __kernel __attribute__((reqd_work_group_size(8, 8, 1)))
